[test] Resolve golden data of size 0
This patch resolves the issue of golden data of size 0.
This happened as a tensor of size 0 was trying to load the data.
As the size was 0, no data was loaded.
Further, this size was used for comparison as well, which led to
faulty testcases being passed as well.

Also batchnormalization layers test were not loading the data.
That is also resolved here.

Resolves #769 #1172
See Also #1180

**Self evaluation:**
1. Build test: [x]Passed [ ]Failed [ ]Skipped
2. Run test: [x]Passed [ ]Failed [ ]Skipped

Signed-off-by: Parichay Kapoor <pk.kapoor@samsung.com>